Amend isbn to big integer to fix load data bug

+++ books/models.py
@@ -19,7 +19,7 @@ class Book(models.Model):
     title = models.CharField(max_length=254)
     summary = models.TextField(max_length=500)
     author = models.CharField(max_length=254)
-    isbn = models.IntegerField()
+    isbn = models.BigIntegerField()
     price = models.DecimalField(max_digits=5, decimal_places=2)
     cover_url = models.URLField(max_length=1024, null=True, blank=True)
     cover = models.ImageField(null=True, blank=True)
